requirements.txt
supabase
openai
python-dotenv
pinecone
